backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
requests==2.31.0
google-api-python-client==2.97.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
